✨ Texture 및 PlayerObject 클래스에서 double을 float로 변경하여 일관성 및 성능 개선

+++ SDL2_SuikaGame/Source/Engine/EngineTypes/FTexture2D.cpp
@@ -10,7 +10,7 @@ FTexture2D::FTexture2D(
 {
     int w, h;
     SDL_QueryTexture(texture, nullptr, nullptr, &w, &h);
-    tex_size = { static_cast<double>(w), static_cast<double>(h) };
+    tex_size = { static_cast<float>(w), static_cast<float>(h) };
 }
 
 FTexture2D::FTexture2D(
