+++ SDL2_SuikaGame/Source/Engine/EngineTypes/Vector2D.cpp
@@ -6,40 +6,40 @@ Vector2D Vector2D::Zero = Vector2D(0.0, 0.0);
 Vector2D Vector2D::One = Vector2D(1.0, 1.0);
 
 Vector2D::Vector2D()
-    : X(0.0)
-    , Y(0.0)
+    : X(0.0f)
+    , Y(0.0f)
 {
 }
 
-Vector2D::Vector2D(double value)
+Vector2D::Vector2D(float value)
     : X(value)
     , Y(value)
 {
 }
 
-Vector2D::Vector2D(double x, double y)
+Vector2D::Vector2D(float x, float y)
     : X(x)
     , Y(y)
 {
 }
 
 
-double Vector2D::DotProduct(const Vector2D& lhs, const Vector2D& rhs)
+float Vector2D::DotProduct(const Vector2D& lhs, const Vector2D& rhs)
 {
     return lhs | rhs;
 }
 
-double Vector2D::CrossProduct(const Vector2D& lhs, const Vector2D& rhs)
+float Vector2D::CrossProduct(const Vector2D& lhs, const Vector2D& rhs)
 {
     return lhs ^ rhs;
 }
 
-double Vector2D::DistanceSquared(const Vector2D& lhs, const Vector2D& rhs)
+float Vector2D::DistanceSquared(const Vector2D& lhs, const Vector2D& rhs)
 {
     return Math::Square(rhs.X - lhs.X) + Math::Square(rhs.Y - lhs.Y);
 }
 
-double Vector2D::Distance(const Vector2D& lhs, const Vector2D& rhs)
+float Vector2D::Distance(const Vector2D& lhs, const Vector2D& rhs)
 {
     return Math::Sqrt(DistanceSquared(lhs, rhs));
 }
@@ -52,7 +52,7 @@ Vector2D Vector2D::Clamp(const Vector2D& value, const Vector2D& min, const Vecto
     );
 }
 
-Vector2D Vector2D::Clamp(const Vector2D& value, double min, double max)
+Vector2D Vector2D::Clamp(const Vector2D& value, float min, float max)
 {
     return Vector2D(
         Math::Clamp(value.X, min, max),
@@ -60,33 +60,33 @@ Vector2D Vector2D::Clamp(const Vector2D& value, double min, double max)
     );
 }
 
-double Vector2D::Length() const
+float Vector2D::Length() const
 {
     return sqrt(X * X + Y * Y);
 }
 
-double Vector2D::LengthSquared() const
+float Vector2D::LengthSquared() const
 {
     return X * X + Y * Y;
 }
 
 Vector2D Vector2D::Normalize() const
 {
-    double length = Length();
+    float length = Length();
     return Vector2D(X / length, Y / length);
 }
 
-double Vector2D::Dot(const Vector2D& rhs) const
+float Vector2D::Dot(const Vector2D& rhs) const
 {
     return *this | rhs;
 }
 
-double Vector2D::Cross(const Vector2D& rhs) const
+float Vector2D::Cross(const Vector2D& rhs) const
 {
     return *this ^ rhs;
 }
 
-bool Vector2D::IsNearlyZero(double tolerance) const
+bool Vector2D::IsNearlyZero(float tolerance) const
 {
     return Math::Abs(X) <= tolerance && Math::Abs(Y) <= tolerance;
 }
@@ -96,7 +96,7 @@ bool Vector2D::IsZero() const
     return X == 0.0 && Y == 0.0;
 }
 
-bool Vector2D::Equals(const Vector2D& rhs, double tolerance) const
+bool Vector2D::Equals(const Vector2D& rhs, float tolerance) const
 {
     return Math::Abs(X - rhs.X) <= tolerance && Math::Abs(Y - rhs.Y) <= tolerance;
 }
@@ -106,7 +106,7 @@ Vector2D Vector2D::operator+(const Vector2D& rhs) const
     return Vector2D(X + rhs.X, Y + rhs.Y);
 }
 
-Vector2D Vector2D::operator+(double value) const
+Vector2D Vector2D::operator+(float value) const
 {
     return Vector2D(X + value, Y + value);
 }
@@ -118,7 +118,7 @@ Vector2D& Vector2D::operator+=(const Vector2D& rhs)
     return *this;
 }
 
-Vector2D& Vector2D::operator+=(double value)
+Vector2D& Vector2D::operator+=(float value)
 {
     X += value;
     Y += value;
@@ -130,7 +130,7 @@ Vector2D Vector2D::operator-(const Vector2D& rhs) const
     return Vector2D(X - rhs.X, Y - rhs.Y);
 }
 
-Vector2D Vector2D::operator-(double value) const
+Vector2D Vector2D::operator-(float value) const
 {
     return Vector2D(X - value, Y - value);
 }
@@ -142,7 +142,7 @@ Vector2D& Vector2D::operator-=(const Vector2D& rhs)
     return *this;
 }
 
-Vector2D& Vector2D::operator-=(double value)
+Vector2D& Vector2D::operator-=(float value)
 {
     X -= value;
     Y -= value;
@@ -154,7 +154,7 @@ Vector2D Vector2D::operator*(const Vector2D& rhs) const
     return Vector2D(X * rhs.X, Y * rhs.Y);
 }
 
-Vector2D Vector2D::operator*(double value) const
+Vector2D Vector2D::operator*(float value) const
 {
     return Vector2D(X * value, Y * value);
 }
@@ -166,7 +166,7 @@ Vector2D& Vector2D::operator*=(const Vector2D& rhs)
     return *this;
 }
 
-Vector2D& Vector2D::operator*=(double value)
+Vector2D& Vector2D::operator*=(float value)
 {
     X *= value;
     Y *= value;
@@ -178,7 +178,7 @@ Vector2D Vector2D::operator/(const Vector2D& rhs) const
     return Vector2D(X / rhs.X, Y / rhs.Y);
 }
 
-Vector2D Vector2D::operator/(double value) const
+Vector2D Vector2D::operator/(float value) const
 {
     return Vector2D(X / value, Y / value);
 }
@@ -190,7 +190,7 @@ Vector2D& Vector2D::operator/=(const Vector2D& rhs)
     return *this;
 }
 
-Vector2D& Vector2D::operator/=(double value)
+Vector2D& Vector2D::operator/=(float value)
 {
     X /= value;
     Y /= value;
@@ -209,8 +209,8 @@ bool Vector2D::operator!=(const Vector2D& rhs) const
 
 std::partial_ordering Vector2D::operator<=>(const Vector2D& rhs) const
 {
-    double lhs_length = Length();
-    double rhs_length = rhs.Length();
+    float lhs_length = Length();
+    float rhs_length = rhs.Length();
 
     if (lhs_length < rhs_length)
     {
@@ -230,12 +230,12 @@ Vector2D Vector2D::operator-() const
     return Vector2D(-X, -Y);
 }
 
-double Vector2D::operator|(const Vector2D& rhs) const
+float Vector2D::operator|(const Vector2D& rhs) const
 {
     return X * rhs.X + Y * rhs.Y;
 }
 
-double Vector2D::operator^(const Vector2D& rhs) const
+float Vector2D::operator^(const Vector2D& rhs) const
 {
     return X * rhs.Y - Y * rhs.X;
 }
